Moved tests under demeter/
This allowed the use of a test helper throughout the tests.
