+++ demeter/tests/app/test_api.py
@@ -20,22 +20,18 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-import uuid
-
 from ddt import data
 from ddt import ddt
 from ddt import unpack
 from flask import json
 import unittest2 as unittest
 
-import demeter.app.api as app
+from demeter.app import api as app
+from demeter.tests import helper
 
 
 @ddt
 class TestApi(unittest.TestCase):
-    def namespace_data(name=str(uuid.uuid4()), cidr='198.51.100.0/24'):
-        return (name, cidr)
-
     def setUp(self):
         self._app = app.app.test_client()
 
@@ -47,7 +43,7 @@ class TestApi(unittest.TestCase):
         self.assertEquals(True, data['success'])
 
     @unpack
-    @data(namespace_data())
+    @data(helper.namespace_data())
     def test_app_create_namespace(self, ns_name, cidr):
         data = json.dumps({"cidr": cidr})
         url = '/v1.0/namespace/{0}'.format(ns_name)
@@ -63,7 +59,7 @@ class TestApi(unittest.TestCase):
         self._app.delete('/v1.0/namespace/api-test-1')
 
     @unpack
-    @data(namespace_data())
+    @data(helper.namespace_data())
     def test_app_create_namespace_returns_409_when_exists(self, ns_name, cidr):
         data = json.dumps({"cidr": cidr})
         url = '/v1.0/namespace/{0}'.format(ns_name)
